requirements.txt
dj-database-url==0.4.0
Django==1.9.2
django-model-utils==2.4
django-suit==0.2.16
django-suit-redactor==0.0.3
gunicorn==19.4.5
opbeat==3.2.1
Unipath==1.1
whitenoise==2.0.6
psycopg2==2.6.1
jsonfield==1.0.3